nfc(major-refactor): File structure changes
