+++ main.hpp
@@ -2,8 +2,8 @@
 
 #include <iostream>
 #include "src/NeuralNet/Network.hpp"
-#include "src/NeuralNet/Dense.hpp"
-#include "src/NeuralNet/interfaces/Layer.hpp"
+#include "src/NeuralNet/layers/Dense.hpp"
+#include "src/NeuralNet/layers/Layer.hpp"
 #include "src/NeuralNet/utils/Functions.hpp"
 #include "src/NeuralNet/optimizers/optimizers.hpp"
 #include <fstream>
