+++ src/CMakeLists.txt
@@ -1,6 +1,5 @@
 
 set(SRC_FILES 
-  ${NETWORK_DIR}/Flatten.cpp
   ${NETWORK_DIR}/Network.cpp
 )
