+++ src/NeuralNet/optimizers/SGD.hpp
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "interfaces/Optimizer.hpp"
+#include "Optimizer.hpp"
 
 namespace NeuralNet
 {
